enh: use untrusted egress proxy for slackbot downloads
