+++ connectors/src/connectors/webcrawler/lib/utils.ts
@@ -3,40 +3,17 @@ import { Err, Ok } from "@dust-tt/client";
 import { hash as blake3 } from "blake3";
 import { NonRetryableError } from "crawlee";
 import dns from "dns";
-import type {
-  RequestInfo as UndiciRequestInfo,
-  RequestInit as UndiciRequestInit,
-} from "undici";
-import { fetch as undiciFetch, ProxyAgent } from "undici";
 
-import { apiConfig } from "@connectors/lib/api/config";
 import type {
   WebCrawlerFolder,
   WebCrawlerPage,
 } from "@connectors/lib/models/webcrawler";
+import { createProxyAwareFetch } from "@connectors/lib/proxy";
 import type { ContentNodeType } from "@connectors/types";
 import { WEBCRAWLER_MAX_DEPTH } from "@connectors/types";
 
 const MAX_REDIRECTS = 20;
 const REDIRECT_STATUSES = new Set([301, 302, 303, 307, 308]);
-
-// Create a fetch function with proxy support if configured
-function createProxyAwareFetch() {
-  const proxyHost = apiConfig.getUntrustedEgressProxyHost();
-  const proxyPort = apiConfig.getUntrustedEgressProxyPort();
-
-  if (proxyHost && proxyPort) {
-    const proxyUrl = `http://${proxyHost}:${proxyPort}`;
-    const dispatcher = new ProxyAgent(proxyUrl);
-
-    return (input: UndiciRequestInfo, init?: UndiciRequestInit) => {
-      return undiciFetch(input, { ...init, dispatcher });
-    };
-  }
-
-  // If no proxy configured, use standard fetch
-  return fetch;
-}
 
 export type WebCrawlerErrorName =
   | "PRIVATE_IP"
